+++ lib/src/analyzers/lint_analyzer/rules/rules_list/ban_name/visitor.dart
@@ -7,26 +7,26 @@ class _Visitor extends GeneralizingAstVisitor<void> {
 
   Iterable<_NodeInfo> get nodes => _nodes;
 
-  final _nodeBreadcrumb = <String, AstNode>{};
+  final _visitedNodes = <AstNode>{};
 
   _Visitor(List<_BanNameConfigEntry> entries)
       : _entryMap = Map.fromEntries(entries.map((e) => MapEntry(e.ident, e)));
 
   @override
   void visitSimpleIdentifier(SimpleIdentifier node) {
-    _visitIdent(node, node.name);
+    _visitIdent(node, node);
   }
 
   @override
   void visitPrefixedIdentifier(PrefixedIdentifier node) {
-    _visitIdent(node, node.identifier.name);
-    _visitIdent(node, node.prefix.name);
+    _visitIdent(node, node.identifier);
+    _visitIdent(node, node.prefix);
   }
 
   @override
   void visitLibraryIdentifier(LibraryIdentifier node) {
     for (final component in node.components) {
-      _visitIdent(node, component.name);
+      _visitIdent(node, component);
     }
   }
 
@@ -34,43 +34,54 @@ class _Visitor extends GeneralizingAstVisitor<void> {
   void visitDeclaration(Declaration node) {
     final name = node.declaredElement?.displayName;
     if (name != null) {
-      _visitIdent(node, name);
+      _checkBannedName(node, name);
     }
 
     super.visitDeclaration(node);
   }
 
-  void _visitIdent(AstNode node, String name) {
-    final prevNode =
-        _nodeBreadcrumb.isNotEmpty ? _nodeBreadcrumb.values.last : null;
-    if (node.offset - 1 == prevNode?.end) {
-      _nodeBreadcrumb.addAll({name: node});
-    } else {
-      _nodeBreadcrumb.clear();
+  void _visitIdent(AstNode node, SimpleIdentifier name) {
+    _checkBannedName(name, name.name);
+
+    if (node is PrefixedIdentifier) {
+      _checkBannedName(node, node.name);
     }
 
-    if (_nodeBreadcrumb.isEmpty) {
-      _nodeBreadcrumb.addAll({name: node});
+    _traverseParents(name.parent);
+  }
+
+  void _traverseParents(AstNode? node) {
+    if (node is MethodInvocation) {
+      _checkBannedName(
+        node,
+        '${node.realTarget}.${node.methodName}',
+      );
+      _traverseParents(node.parent);
     }
 
-    if (_entryMap.containsKey(name)) {
+    if (node is PropertyAccess) {
+      _checkBannedName(
+        node,
+        '${node.realTarget}.${node.propertyName}',
+      );
+      _traverseParents(node.realTarget);
+    }
+
+    if (node is ConstructorName) {
+      _traverseParents(node.parent);
+    }
+
+    if (node is InstanceCreationExpression) {
+      _checkBannedName(node, node.constructorName.toString());
+    }
+  }
+
+  void _checkBannedName(AstNode node, String name) {
+    if (_entryMap.containsKey(name) && !_visitedNodes.contains(node)) {
+      _visitedNodes.add(node);
       _nodes.add(_NodeInfo(
         node,
-        fullName: name,
         message: '${_entryMap[name]!.description} ($name is banned)',
-      ));
-
-      return;
-    }
-
-    final breadcrumbString = _nodeBreadcrumb.keys.join('.');
-    if (_entryMap.containsKey(breadcrumbString)) {
-      _nodes.add(_NodeInfo(
-        _nodeBreadcrumb.values.first,
-        fullName: breadcrumbString,
-        message:
-            '${_entryMap[breadcrumbString]!.description} ($breadcrumbString is banned)',
-        endNode: _nodeBreadcrumb.values.last,
       ));
     }
   }
@@ -78,14 +89,10 @@ class _Visitor extends GeneralizingAstVisitor<void> {
 
 class _NodeInfo {
   final AstNode node;
-  final String fullName;
   final String message;
-  final AstNode? endNode;
 
   _NodeInfo(
     this.node, {
-    required this.fullName,
     required this.message,
-    this.endNode,
   });
 }